feat: replaced guild with dao and extracted common hooks
replaced guild references with DAO, and hooks used by various implementations are stored in commons folder
