+++ src/stores/modules/common/fetchers/useSnapshotId.ts
